+++ Changes.md
@@ -2,9 +2,11 @@
 
 ## [1.3](https://github.com/folio-org/Net-Z3950-FOLIO/tree/v1.3) (IN PROGRESS)
 
+* Make robust when dealing with MARC fields that have no subfields.
 * Add dependency on `source-storage-source-records` interface.
 * When substituting environment variables in the configuration file, recognise the bash-like fallback syntax `${NAME-VALUE}`, which uses the value of the environment variable `NAME` when defined, falling back to the constant value `VALUE` otherwise. This allows the configuration to include default values which can be overridden with environment variables.
-* TODO: more polite rejection of searches with no access-point specified.
+* Use `default` configured index when no Z39.50 access-point is specified for a search.
+* Support Z39.50 sorting.
 * TODO: documentation of server capabilities.
 
 ## [1.2](https://github.com/folio-org/Net-Z3950-FOLIO/tree/v1.2) (Fri Sep 18 10:15:11 BST 2020)
@@ -25,6 +27,5 @@
 ## To do
 
 * Determine FOLIO tenant from database name (and postpone initialisation and authentication until we know that).
-* Support Z39.50 sorting.
 * Automatic generation of MARC records (will need a non-trivial version of `etc/folio2marcxml.xsl`).